hardware/poc/CamTest2/utility/requirements.txt
backcall==0.2.0
decorator==5.0.7
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
numpy==1.20.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.8.1
pyserial==3.5
traitlets==5.0.5
wcwidth==0.2.5
